# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Tax Intelligence Engine API - Pakistan-focused AI Tax Software
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault/documents:
    get:
      operationId: listVaultDocuments
      tags: [vault]
      summary: List knowledge vault documents (Super Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: taxType
          in: query
          schema:
            type: string
        - name: docType
          in: query
          schema:
            type: string
        - name: jurisdiction
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Vault documents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocumentListResponse"
    post:
      operationId: uploadVaultDocument
      tags: [vault]
      summary: Upload document to knowledge vault (Super Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/VaultUploadRequest"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
  /vault/documents/{id}:
    get:
      operationId: getVaultDocument
      tags: [vault]
      summary: Get vault document details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vault document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
    put:
      operationId: updateVaultDocument
      tags: [vault]
      summary: Update vault document metadata (Super Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaultDocumentRequest"
      responses:
        "200":
          description: Document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
    delete:
      operationId: deleteVaultDocument
      tags: [vault]
      summary: Delete vault document (Super Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /vault/documents/{id}/activate:
    patch:
      operationId: activateVaultDocument
      tags: [vault]
      summary: Activate vault document (Super Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
  /vault/documents/{id}/deactivate:
    patch:
      operationId: deactivateVaultDocument
      tags: [vault]
      summary: Deactivate vault document (Super Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDocument"
  /vault/documents/{id}/versions:
    get:
      operationId: getVaultDocumentVersions
      tags: [vault]
      summary: Get version history for a vault document
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultVersionListResponse"
  /vault/search:
    post:
      operationId: searchVault
      tags: [vault]
      summary: Search vault documents for AI retrieval
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultSearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSearchResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    VaultDocument:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        docType:
          type: string
        taxType:
          type: string
        jurisdiction:
          type: string
        lawSection:
          type: string
        issueDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
        priority:
          type: integer
        versionNo:
          type: integer
        tags:
          type: array
          items:
            type: string
        summary:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        uploadedBy:
          type: string
        indexedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - docType
        - status
    VaultUploadRequest:
      type: object
      properties:
        title:
          type: string
        docType:
          type: string
        taxType:
          type: string
        jurisdiction:
          type: string
        lawSection:
          type: string
        issueDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        priority:
          type: integer
        tags:
          type: string
        file:
          type: string
          format: binary
      required:
        - title
        - docType
        - file
    UpdateVaultDocumentRequest:
      type: object
      properties:
        title:
          type: string
        docType:
          type: string
        taxType:
          type: string
        jurisdiction:
          type: string
        lawSection:
          type: string
        issueDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        priority:
          type: integer
        tags:
          type: array
          items:
            type: string
        summary:
          type: string
    VaultDocumentListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/VaultDocument"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - documents
        - total
    VaultVersion:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        versionNo:
          type: integer
        fileName:
          type: string
        uploadedBy:
          type: string
        changeNote:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - versionNo
    VaultVersionListResponse:
      type: object
      properties:
        versions:
          type: array
          items:
            $ref: "#/components/schemas/VaultVersion"
        total:
          type: integer
      required:
        - versions
        - total
    VaultSearchRequest:
      type: object
      properties:
        query:
          type: string
        taxType:
          type: string
        docType:
          type: string
        jurisdiction:
          type: string
        limit:
          type: integer
          default: 5
      required:
        - query
    VaultSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              document:
                $ref: "#/components/schemas/VaultDocument"
              relevanceScore:
                type: number
              excerpt:
                type: string
              matchedSections:
                type: array
                items:
                  type: string
        totalFound:
          type: integer
      required:
        - results
        - totalFound